Refactoring al Eliminar Prematricula, agregando algunas validaciones.

+++ BackEnd/PreMatricula/Aplicacion/Service/Eliminar/EliminarPreMatriculaService.cs
@@ -4,17 +4,14 @@ using BackEnd.Estudiante.Aplicacion.Services.Eliminar;
 using BackEnd.PreMatricula.Aplicacion.Request;
 using BackEnd.Responsable.Aplicacion.Request;
 using BackEnd.Responsable.Aplicacion.Services.Eliminar;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace BackEnd.PreMatricula.Aplicacion.Service.Eliminar
 {
     public class EliminarPreMatriculaService
     {
         readonly IUnitOfWork _unitOfWork;
-        public EliminarEstudianteService estudianteService;
-        public EliminarResponsableService responsableService;
+        private readonly EliminarEstudianteService estudianteService;
+        private readonly EliminarResponsableService responsableService;
         public EliminarPreMatriculaService(IUnitOfWork unitOfWork)
         {
             _unitOfWork = unitOfWork;
@@ -24,25 +21,27 @@ namespace BackEnd.PreMatricula.Aplicacion.Service.Eliminar
         public EliminarPreMatriculaResponse Ejecutar(EliminarPreMatriculaRequest request)
         {
             Dominio.PreMatricula prematricula = _unitOfWork.PreMatriculaServiceRepository.FindFirstOrDefault(t => t.IdUsuario == request.UsuarioId);
+            if (prematricula == null)
+            {
+                return new EliminarPreMatriculaResponse($"PreMatricula no existe");
+            }
+            if (prematricula.Estado == "Confirmado")
+            {
+                return new EliminarPreMatriculaResponse($"Prematricula confirmada, primero elimine la matrícula");
+            }
             EliminarEstudianteRequest estudianteRequest = new EliminarEstudianteRequest();
             estudianteRequest.IdUsuario = request.UsuarioId;
             EliminarResponsableRequest responsableRequest = new EliminarResponsableRequest();
             responsableRequest.UsuarioId = request.UsuarioId;
             var respuestaE = estudianteService.Ejecutar(estudianteRequest);
-            for (int i = 1; i <= 3; i++)
-            {
-                var respuesta = responsableService.Ejecutar(responsableRequest);
-            }
-            if (prematricula == null)
-            {
-                return new EliminarPreMatriculaResponse() { Message = $"PreMatricula no existe" };
-            }
-            else
+            var respuesta = responsableService.Ejecutar(responsableRequest);
+            if (respuesta.isOk() && respuestaE.isOk())
             {
                 _unitOfWork.PreMatriculaServiceRepository.Delete(prematricula);
                 _unitOfWork.Commit();
-                return new EliminarPreMatriculaResponse() { Message = $"PreMatricula Eliminado Exitosamente" };
+                return new EliminarPreMatriculaResponse($"PreMatricula Eliminado Exitosamente");
             }
+            return new EliminarPreMatriculaResponse($"PreMatricula no se Eliminó");
         }
     }
 }
